requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
edge-tts>=6.1.0
pyttsx3>=2.90
pydub>=0.25.1
srt>=3.5.3
requests>=2.31.0
pathlib
gTTS==2.5.1